Repurpose takingExplicitActionsOf test for AclAclLine
